+++ mne/io/fiff/raw.py
@@ -22,7 +22,7 @@ from ..base import (BaseRaw, _RawShell, _check_raw_compatibility,
                     _check_maxshield)
 from ..utils import _mult_cal_one
 
-from ...annotations import Annotations, _combine_annotations, _onset_to_seconds
+from ...annotations import Annotations, _combine_annotations, _sync_onset
 
 from ...event import AcqParserFIF
 from ...utils import check_fname, logger, verbose, warn
@@ -134,7 +134,7 @@ class Raw(BaseRaw):
                     start = skip['first'] - first_samp + offset
                     stop = skip['last'] - first_samp - 1 + offset
                     self.annotations.append(
-                        _onset_to_seconds(self, start / self.info['sfreq']),
+                        _sync_onset(self, start / self.info['sfreq']),
                         (stop - start) / self.info['sfreq'], 'BAD_ACQ_SKIP')
         if preload:
             self._preload_data(preload)