+++ mne/tests/test_annotations.py
@@ -3,8 +3,9 @@
 # License: BSD 3 clause
 
 from datetime import datetime
-from nose.tools import assert_raises
-from numpy.testing import assert_array_equal, assert_array_almost_equal
+from nose.tools import assert_raises, assert_true
+from numpy.testing import (assert_equal, assert_array_equal,
+                           assert_array_almost_equal)
 import os.path as op
 
 import numpy as np
@@ -84,5 +85,33 @@ def test_annotations():
 
     assert_array_almost_equal(raw.annotations.onset, [45., 2. + last_time],
                               decimal=2)
+
+
+@testing.requires_testing_data
+def test_raw_reject():
+    """Test raw data getter with annotation reject."""
+    info = create_info(['a', 'b', 'c', 'd', 'e'], 100, ch_types='eeg')
+    raw = RawArray(np.ones((5, 15000)), info)
+    raw.annotations = Annotations([2, 100, 105, 148], [2, 8, 5, 8], 'BAD')
+    data = raw.get_data([0, 1, 3, 4], 100, 11200, 'omit')
+    assert_array_equal(data.shape, (4, 9900))
+
+    # with orig_time and complete overlap
+    raw = read_raw_fif(fif_fname)
+    raw.annotations = Annotations([44, 47, 48], [1, 3, 1], 'BAD',
+                                  raw.info['meas_date'])
+    data, times = raw.get_data(range(10), 0, 6000, 'omit', True)
+    assert_array_equal(data.shape, (10, 4799))
+    assert_equal(times[-1], raw.times[5999])
+    assert_array_equal(data[:, -100:], raw[:10, 5900:6000][0])
+
+    data, times = raw.get_data(range(10), 0, 6000, 'NaN', True)
+    assert_array_equal(data.shape, (10, 6000))
+    assert_equal(times[-1], raw.times[5999])
+    assert_true(np.isnan(data[:, 313:613]).all())  # 1s -2s
+    assert_true(not np.isnan(data[:, 614].any()))
+    assert_array_equal(data[:, -100:], raw[:10, 5900:6000][0])
+    assert_array_equal(raw.get_data(), raw[:][0])
+
 
 run_tests_if_main()